feat: moves GitClone rule to its own module
